Changes to structure - animate and fetch seprate files -working
