+++ src/layouts/Layout.astro
@@ -114,4 +114,5 @@ const { title } = Astro.props;
 </style>
 <script>
   import "../scripts/fetch.js";
+  import "../scripts/animation.js";
 </script>
