+++ src/pages/index.astro
@@ -30,25 +30,25 @@ import ButtonComponent from "../components/ButtonComponent.astro";
 
 			<IconWithText
           	iconSrc="../../public/icons/pasport.svg"
-          	mainText="Temperament data here"
           	subText="/Temperament"
+            id=""
         	/>
 	-->
       <div class="text-container">
         <IconWithText
           iconSrc="../../public/icons/pasport.svg"
-          mainText="Temperament data here"
           subText="/Temperament"
+          id="fetchContentOne"
         />
         <IconWithText
           iconSrc="../../public/icons/box.svg"
-          mainText="Origin data here"
           subText="/Origin"
+          id="fetchContentTwo"
         />
         <IconWithText
           iconSrc="../../public/icons/energy.svg"
-          mainText="Energy level here"
           subText="/Energy level"
+          id="fetchContentThree"
         />
       </div>
     </div>
